requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\tanya\Desktop\python\2 semester\extra tasks\portfolio_messenger\server.py: 2
Flask == 1.1.2

# C:\Users\tanya\Desktop\python\2 semester\extra tasks\portfolio_messenger\clientui.py: 12
# C:\Users\tanya\Desktop\python\2 semester\extra tasks\portfolio_messenger\launch_messenger.py: 3
PyQt5_sip == 12.7.2

# C:\Users\tanya\Desktop\python\2 semester\extra tasks\portfolio_messenger\launch_messenger.py: 2
# C:\Users\tanya\Desktop\python\2 semester\extra tasks\portfolio_messenger\server.py: 4
# C:\Users\tanya\Desktop\python\2 semester\extra tasks\portfolio_messenger\utils\client.py: 1
# C:\Users\tanya\Desktop\python\2 semester\extra tasks\portfolio_messenger\utils\getter.py: 3
requests == 2.23.0
